Add Private DNS warning
Since this is a DNS based hosts blocker, we need to be the sole provider of DNS requests. Since Android 10 and later, all DNS requests will be served through the configured Private DNS server as stated by the user setting or the currently connected network. Only when this is disabled, we can function as normal. As a result, we have to show a blocking dialog to the user to clarify what may be going wrong.