Drop always-on VPN support
I can't quite figure out how to get this working properly and it's honestly less effort to just disable this and achieve the exact functionality that I'm looking for.
Before, this VPN service would be busted because we're not handling the "Block connections without VPN" setting correctly. Now, it works as intended.