Potential fix for first time startup failure
Sometimes during the VPN's first start, it can get into an infinite reconnecting loop because it never successfully prepares the file descriptor for the connection. The documentation says that this is because the service is not prepared. So here, I added an additional check to make sure that the service is prepared and to stop the VPN immediately if we are not.