Write logcat to file instead of share sheet
For whatever reason, it appears as though there are limits on the amount of data that you can put into a share sheet request. I couldn't find anything online about this, but I keep seeing this issue pop up during testing. So instead, we just open a buffered read/write to a file that the user creates.