Create option to add a single host from Hosts screen
Previously this was only available as a result of making an exception in the block log, but now you can access this from a new expandable floating action button component in the hosts screen.

This also changes the language around adding an "exception." This is just adding a single hostname so it was renamed.